requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
sqlparse==0.5.2
typing_extensions==4.12.2
